requirements.txt
requests
beautifulsoup4
psycopg2-binary
APScheduler
slack_sdk
python-dotenv